Fixing warnings
Such a mess..  longs to ints etc...  yuck

+++ src/files.h
@@ -70,7 +70,7 @@ public:
 	virtual long Seek (long offset, int origin);
 	virtual long Read (void *buffer, long len);
 	virtual char *Gets(char *strbuf, int len);
-	long GetLength () const { return Length; }
+	int GetLength () const { return Length; }
 
 	// If you use the underlying FILE without going through this class,
 	// you must call ResetFilePtr() before using this class again.
@@ -119,7 +119,7 @@ protected:
 	char *GetsFromBuffer(const char * bufptr, char *strbuf, int len);
 
 	FILE *File;
-	long Length;
+	int Length;
 	long StartPos;
 	long FilePos;
